pyproject.toml
[project]
name = "audio-stream-server"
version = "0.1.0"
description = "Stream MP3 from Youtube videos over HTTP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "jinja2",
    "openai>=1.0.0",
    "google-genai",
    "httpx",
    "python-dotenv",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]
dev = [
    "mypy>=1.8.0",
    "black>=24.0.0",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = false
warn_redundant_casts = true
warn_no_return = false
strict_optional = false

# Gradual typing - start lenient
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_any_generics = false
no_implicit_optional = false

# Ignore common patterns for now
[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

